requirements.txt
numpy
pymysql
pandas
scipy
TA-Lib
technical
openpyxl
cryptography
discord.py
python-dotenv
tabulate
requests
wheel>=0.36.2